requirements.txt
flask
Flask-SQLAlchemy
psycopg2-binary
PyGithub===1.54.1
gitpython===3.1.11
requests
PyQuery
python-dotenv
flask-cors